refactor(github): move GitHub App client to pkg/github (#6839)
* refactor(github): move GitHub App client to pkg/github

The GitHub App client is shared infrastructure used by svc/ctrl (deploy, webhooks, status) and svc/api (app repository connections). Moving it under pkg/github removes the cross-service import of svc/ctrl/worker internals and gives every service a clean, standalone import.

* style: fmt

---------
